requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.10.3
scikit-learn==1.5.2
numpy==1.26.4
groq==0.12.0
